Fix emojis being rendered multiple times on glowing signs Refactor mixin package structure Add MixinHelpers.java
