+++ src/main/java/io/github/aratakileo/emogg/mixin/MixinHelpers.java
@@ -1,0 +1,8 @@
+package io.github.aratakileo.emogg.mixin;
+
+public class MixinHelpers {
+    // FontMixin.preGlowOutlineRender
+    // FontMixin.postGlowOutlineRender
+    // StringRenderOutputMixin.noShadowAndGlowOutlineForEmojis
+    public static boolean shouldSkipEmojiGlyphRender = false;
+}
