+++ src/main/java/io/github/aratakileo/emogg/mixin/mixins/gui/MinecraftMixin.java
@@ -1,4 +1,4 @@
-package io.github.aratakileo.emogg.mixin.gui;
+package io.github.aratakileo.emogg.mixin.mixins.gui;
 
 import io.github.aratakileo.emogg.Emogg;
 import io.github.aratakileo.emogg.api.ModrinthApi;
