+++ src/main/resources/emogg.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "io.github.aratakileo.emogg.mixin",
+  "package": "io.github.aratakileo.emogg.mixin.mixins",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
   ],
